fully implemented backpropagation
I realized there was a massive design error in the way that I was implementing backprop in the previous commit. That being the way that the gradients were being accumulated and averaged into the network struct led to an issue where the gradient being propagated to the lower layers was only a single averaged gradient over the batch, rather than all the gradients for the batch, this made back prop of the lwoer layer infeasible. I have corrected this by creating a master function that backprops the grads for an entire forward pass, accumulates, and averages after all grads are computed and accumulated.