beggining refactoring design of entire project
After taking a short break from this project, I realized that my process of design and implementaiton was carried out too recklessly. This gave me a good intuition for what goes into a project liek this. However, it now requires that I start the entire process over from scratch in order to do it right and simply. For the time being, the og_implementation directory will contain the original gangsta implementaiton as a refference, and the refactored_implementation dir will contain the new implementaiton